requirements.txt
PyMuPDF==1.26.7
pyiceberg[glue]==0.8.1
pyarrow==18.1.0
boto3==1.35.99
python-dotenv==1.0.1
anthropic>=0.76.0
requests>=2.31.0
snowflake-connector-python>=3.6.0
dbt-snowflake>=1.7.0
